Refacored into separate file and improved tests
